Remove nose from validation_cp
