+++ validation/run_validation_tests.py
@@ -1,30 +1,14 @@
+import contextlib
 import os.path as osp
-import sys
-import collections.abc
-collections.Callable = collections.abc.Callable
-
-import nose
-
-
-DEFAULT_TEST_SUITE = [
-    'validation_cp',
-    'validation_efield',
-    'validation_rd',
-]
-
-
-def main():
-    test_dir = osp.dirname(osp.abspath(__file__))
-    test_suite = sys.argv[1:] or DEFAULT_TEST_SUITE
-    for suite in test_suite:
-        nose.run(argv=[
-            __file__,
-            '-s',
-            '--all-modules',
-            '-v',
-            osp.join(test_dir, suite)
-        ])
-
+import unittest
 
 if __name__ == '__main__':
-    main()
+    test_dir = osp.dirname(osp.abspath(__file__))
+    loader = unittest.TestLoader()
+    # Redirect stdout to avoid STEPS prints on module import
+    with contextlib.redirect_stdout(None):
+        top_dir = osp.join(test_dir, '..')
+        suite = loader.discover(test_dir, pattern='test_*.py', top_level_dir=top_dir)
+
+    runner = unittest.TextTestRunner()
+    runner.run(suite)
